+++ exp/simd/simd.go
@@ -3,10 +3,10 @@ package simd
 // Empty file to satisfy the build constraint for non-supported architectures.
 
 const (
-	simdLanes2  = 2
-	simdLanes4  = 4
-	simdLanes8  = 8
-	simdLanes16 = 16
-	simdLanes32 = 32
-	simdLanes64 = 64
+	simdLanes2  = uint(2)
+	simdLanes4  = uint(4)
+	simdLanes8  = uint(8)
+	simdLanes16 = uint(16)
+	simdLanes32 = uint(32)
+	simdLanes64 = uint(64)
 )
